Split workflows (#16)
Previously I had in the readme instructions `Make sure to check "Allow edits by maintainers" on your fork PR`, but I forgot that [this option is disabled for organizations](https://github.com/orgs/community/discussions/5634). (We can't have users create branches directly because then every time a new person submits a PR they'd have to request to become a collaborator in the repo, which will get tiring.) Without the "Allow edits" checkbox, Actions workflows do not have permissions to make commits to the fork (by default; I think a user could enable it in their fork settings). As such, this PR splits workflows into read (check-only) and write (auto-fix). The former runs on PRs, the latter runs on `main` where it has the permissions to do so. This PR also switches from the exiftool CLI to the npm package wrapper for the CLI tool for a bit nicer ergonomics.